Model contains extra field "Location". Search engine is added. User rating is saved on user interaction.
The website requires one additional folder inside `<webapp folder>/static/img/`,
which contains all the images. It must be named` images` and has to mirror the structure
of a folder which Lukas shared.

I forced git to add a few extra files from `templates/` as well as `artwork_metadata/`
to add an augmented CSV file Lukas extracted for me; the CSV file contains
relative paths to all images as the pattern in the CSV file does not necessarily reflect
the location of each file.

The model has also changed slightly as it contains relative paths.

A search engine works and accepts multiple terms, but each term is treated as a logical OR.
It turns out I haven't had to do much on this one because Jonathan almost completed it.

The "rating" saves the user ratings in the database when user interacts with it.