A search engine has been improved and accepts the pattern keyword: arguments
The search bar accepts three keywords: title, artist, and medium, followed by the colon,
for instance, "artist:". Currently, other keywords will cause the query engine to return
the entire database and should not be in the query string. The subqueries can be combined
by "&", a logical AND to find the intersection of two subsets. A space character is always
a logical OR.

Some examples:
artist: Cook
artist: cook & medium: metal
etc.

The new engine requires the new module `pyparsing`.